Pass client trace metadata to pages router clients
